Changed how meshes are loaded (added MeshSubSet)

+++ src/Titan.Graphics/Buffers/IndexBuffer.cs
@@ -3,15 +3,16 @@ using Titan.D3D11.Device;
 
 namespace Titan.Graphics.Buffers
 {
+    // TODO: Make it possible to select u16 or u32 for indices
     internal class IndexBuffer : IIndexBuffer
     {
         private readonly ID3D11Buffer _buffer;
-        private readonly short[] _indices;
+        private readonly int[] _indices;
         public IntPtr NativeHandle => _buffer.Handle;
-        public ref readonly short[] Indicies => ref _indices; // should this me stored in memory? we can just access the buffer in the GPU
+        public ref readonly int[] Indicies => ref _indices; // should this me stored in memory? we can just access the buffer in the GPU
         public uint NumberOfIndices { get; private set; }
 
-        public IndexBuffer(ID3D11Buffer buffer, in short[] indices)
+        public IndexBuffer(ID3D11Buffer buffer, in int[] indices)
         {
             _buffer = buffer ?? throw new ArgumentNullException(nameof(buffer));
             _indices = indices;
