+++ src/Titan.Graphics/Rendering/Class1.cs
@@ -145,7 +145,7 @@ namespace Titan.Graphics.Rendering
 
         public override void Build(IDevice device)
         {
-            _indexBuffer = device.CreateIndexBuffer(new short[] { 0, 3, 1, 3, 2, 1 });
+            _indexBuffer = device.CreateIndexBuffer(new int[] { 0, 3, 1, 3, 2, 1 });
             _vertexBuffer = device.CreateVertexBuffer(CreateCube());
 
             using var pixelShaderBlob = new Blob(device.TEMPORARYCompiler.CompileShaderFromFile(@"F:\Git\GameDev\resources\shaders\BackBufferPixelShader.hlsl", "main", "ps_5_0"));
